Try to add a little more helpful information about tgz file references
I was quite confused about why yarn would not accept my `tgz` file.  It seems like I had to put the `package.json` and other files into a folder inside the `tgz` rather than have them bare at the top level.  This requirement didn't seem to be documented anywhere, so I thought I would try to offer an improvement to the docs here so that others (or my future self) will not be as confused by this.