Docusaurus: Update Docusaurus to 3.7 (#6777)
## What's the problem this PR addresses?

The Docusaurus version we use is not the latest

## How did you fix it?

Upgrades Docusuaurs to 3.7 and start using the new features:

- Also upgrade `docusaurus-plugin-typedoc-api` for compatibility
- `docusaurus-plugin-typedoc-api` is unmaintined now. Maybe we should
make our custom solution eventually?
- Upgrade React to v19 since Docusuaurs 3.7 supports it now
- Also upgrades other libraries for official compatibility with React
19, which, while maybe not strictly necessary (nothing seems to break
without those upgrades), is nice to get out of the way
- Upgrade `xterm` so the website can be built with Node 22
- Bump the Netlify build to Node 22, as Node 18 will be EOL this week.
- Add truncation of blog posts on the main blog feed (recommended by
Docusaurus)
- Add author social links to blog posts (@arcanis I just copied the
socials from your website, expect replacing Twitter with BlueSky and
adding the website itself. Please review if there are changes you'd like
to make)
- Remove the extraneous dependencies that [weren't
removed](https://github.com/yarnpkg/berry/pull/6264#issuecomment-2097246182)
in #6264.
- Now I'm pretty sure `markdown-it-br` and `marked-base-url` is actually
usused
- Even if we are to rebuild the sherlock playground, I feel like there
are better libraries for that then `@codesandbox/sandpack-react`. Even
if we use `@codesandbox/sandpack-react` again we'd probably use the
latest version which is a semver-major bump so there will be churn
either way.
- Minor fixes like fixing backgrounds

I have done some visual comparison testing to verify there aren't
unexpected visual changes (on the pages I checked at least)

----

Questions:
- Docusaurus's perf logger now records memory usage. Maybe we should
turn that on on Netlify?
- Do we want to start migrating to `compressionLevel: 0` (the default
since 4.0)? This PR replaces ~300 deps (out of ~2,100) so I feel like
that's a good time to do so *if* we are to migrate

## Checklist

<!--- Don't worry if you miss something, chores are automatically
tested. -->
<!--- This checklist exists to help you remember doing the chores when
you submit a PR. -->
<!--- Put an `x` in all the boxes that apply. -->
- [x] I have read the [Contributing
Guide](https://yarnpkg.com/advanced/contributing).

<!-- See
https://yarnpkg.com/advanced/contributing#preparing-your-pr-to-be-released
for more details. -->
<!-- Check with `yarn version check` and fix with `yarn version check
-i` -->
- [x] I have set the packages that need to be released for my changes to
be effective.

<!-- The "Testing chores" workflow validates that your PR follows our
guidelines. -->
<!-- If it doesn't pass, click on it to see details as to what your PR
might be missing. -->
- [x] I will check that all automated PR checks pass before the PR gets
reviewed.