fix: correctly install implicit nested types dependencies (#6800)
## What's the problem this PR addresses?
Closes #6442

By default - yarn will automatically add `@types/foo` for a package, if
the package
has a peer dependency on `foo`. This happens in the `normalizePackage`
core,
and exists as packages often don't specify the correct peer
dependencies.

However, given a scenario where:
```json
"dependencies": {
    "peer-deps-implicit-types-conflict": "1.0.0",
    "@types/no-deps": "1.0.0",
}
```
where `peer-deps-implicit-types-conflict` has a dependency on
`@types/no-deps:2.0.0`,
```json
  "dependencies": {
    "@types/no-deps": "2.0.0"
  },
  "peerDependencies": {
    "no-deps": "1.0.0"
  }
```
yarn will only install `@types/no-deps:1.0.0` - even though
`@types/no-deps: 2.0.0` is specified
as a dependency!

## How did you fix it?

The fix is rather trivial and thanks to the helpful comment here -
https://github.com/yarnpkg/berry/issues/6442#issuecomment-2273575924 -
simply don't install the automatic `@types/foo` for a package that has
`foo` as a peer-dependency, if the package already specifies
`@types/foo` as a **dependency**.

There is already logic for this behavior when it comes to `@types/`
already existing in `peerDependencies` or `peerDependenciesMeta` - this
PR just extends that to `dependencies` as well.
...

## Checklist

<!--- Don't worry if you miss something, chores are automatically
tested. -->
<!--- This checklist exists to help you remember doing the chores when
you submit a PR. -->
<!--- Put an `x` in all the boxes that apply. -->
- [x] I have read the [Contributing
Guide](https://yarnpkg.com/advanced/contributing).

<!-- See
https://yarnpkg.com/advanced/contributing#preparing-your-pr-to-be-released
for more details. -->
<!-- Check with `yarn version check` and fix with `yarn version check
-i` -->
- [x] I have set the packages that need to be released for my changes to
be effective.

<!-- The "Testing chores" workflow validates that your PR follows our
guidelines. -->
<!-- If it doesn't pass, click on it to see details as to what your PR
might be missing. -->
- [x] I will check that all automated PR checks pass before the PR gets
reviewed.

---------

Co-authored-by: Maël Nison <nison.mael@gmail.com>