Add documentation for opts.conditions of resolveRequest from pnpapi
Is the `conditions` option for the `resolveRequest` function from the pnpapi public? It's currently not documented anywhere.

I came across this as part of visjs/vis-timeline#1898. `resolveRequest` would not resolve the request properly, complaining that it's not defined in the `exports` conditions of the `package.json`.

Then I looked through the .pnpapi.cjs code and found that `resolveRequest` has a `conditions` option that defaults to `node, require`. After setting it to `import`, it could resolve it successfully.