Fixes publishing archives (#6905)
## What's the problem this PR addresses?

We shipped a change that had the unfortunate effect of breaking
publishing packages. It somehow didn't break our tests, which is
something we'll have to investigate. In the meantime I want to fix it
quick.

## How did you fix it?

Revert the probable cause; it wasn't a very important change, and I
found another to achieve what we intended to do.

## Checklist

<!--- Don't worry if you miss something, chores are automatically
tested. -->
<!--- This checklist exists to help you remember doing the chores when
you submit a PR. -->
<!--- Put an `x` in all the boxes that apply. -->
- [x] I have read the [Contributing
Guide](https://yarnpkg.com/advanced/contributing).

<!-- See
https://yarnpkg.com/advanced/contributing#preparing-your-pr-to-be-released
for more details. -->
<!-- Check with `yarn version check` and fix with `yarn version check
-i` -->
- [x] I have set the packages that need to be released for my changes to
be effective.

<!-- The "Testing chores" workflow validates that your PR follows our
guidelines. -->
<!-- If it doesn't pass, click on it to see details as to what your PR
might be missing. -->
- [x] I will check that all automated PR checks pass before the PR gets
reviewed.